backend/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Document Processing
pypdf2>=3.0.0
python-docx>=1.1.0
pillow>=10.1.0
pytesseract>=0.3.10
pdf2image>=1.16.3

# AI/LLM Integration
groq>=0.4.0
httpx>=0.25.0
aiohttp>=3.9.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Security
cryptography>=41.0.0

# Utilities
python-dotenv>=1.0.0
jinja2>=3.1.0
pandas>=2.1.0
openpyxl>=3.1.0
markdown>=3.5.0
markdown-pdf>=0.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
